backend/requirements.txt
asyncio==3.4.3
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
opensearch-py==2.2.0
pyaes==1.6.1
pyasn1==0.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
Telethon==1.28.5
urllib3==1.26.16
